WIP test on define table
