+++ lib/src/kvs/tests/mod.rs
@@ -1,2 +1,2 @@
-mod thing;
+mod tb;
 pub(crate) mod tx_isolation;
